Se aplicaron algunas mejoras

+++ CinemaWebAPI/Controllers/FuncionesController.cs
@@ -22,17 +22,20 @@ namespace CinemaWebAPI.Controllers
         public async Task<ActionResult<IEnumerable<FuncionDTO>>> GetFunciones()
         {
             var funciones = await _context.Funciones
+                .Include(f => f.Sala) // Cargar la relación con la tabla Sala
+                .Include(f => f.Pelicula) // Cargar la relación con la tabla Película
                 .Select(f => new FuncionDTO
                 {
-                    SalaId = f.SalaId,
+                    SalaNombre = f.Sala.Nombre,
                     Fecha = f.Fecha,
                     Horario = f.Horario,
-                    PeliculaId = f.PeliculaId
+                    PeliculaTitulo = f.Pelicula.Titulo
                 })
                 .ToListAsync();
 
             return funciones;
         }
+
 
         // GET: api/Funciones/5
         [HttpGet("{id}")]
@@ -44,11 +47,17 @@ namespace CinemaWebAPI.Controllers
             {
                 return NotFound();
             }
+            var sala = await _context.Salas.FindAsync(funcion.SalaId); // Obtén la sala correspondiente
+            var pelicula = await _context.Peliculas.FindAsync(funcion.PeliculaId);
 
             var funcionDTO = new FuncionDTO
             {
-                FuncionId = funcion.FuncionId,
-                // Mapea otros campos de Funcion a FuncionDTO según tus necesidades.
+                SalaId = funcion.SalaId,
+                SalaNombre = sala != null ? sala.Nombre : "No existe la sala",
+                Fecha = funcion.Fecha,
+                Horario = funcion.Horario,
+                PeliculaId = funcion.PeliculaId,
+                PeliculaTitulo = pelicula != null ? pelicula.Titulo : "No existe la pelicula"
             };
 
             return funcionDTO;
@@ -100,8 +109,9 @@ namespace CinemaWebAPI.Controllers
             {
                 return NotFound(); // La función con el ID proporcionado no existe.
             }
-
-            int cantidadDisponible = funcion.Sala.Capacidad - await _context.Tickets.CountAsync(t => t.FuncionId == id);
+            var sala = await _context.Salas.FindAsync(funcion.SalaId); // Obtén la sala correspondiente
+            int cantidadTickets = await _context.Tickets.CountAsync(t => t.FuncionId == id);
+            int cantidadDisponible = sala.Capacidad - cantidadTickets;
 
             var cantidadTicketsResponse = new CantidadTicketsResponse
             {
@@ -110,43 +120,68 @@ namespace CinemaWebAPI.Controllers
 
             return cantidadTicketsResponse;
         }
-        // POST: api/Funciones
+
         [HttpPost]
         public async Task<ActionResult<FuncionDTO>> PostFuncion(FuncionDTO funcionDTO)
         {
-            if (funcionDTO == null)
-            {
-                return BadRequest("El objeto funcionDTO es nulo.");
-            }
-
-            // Mapea los campos de funcionDTO a la entidad Funcion.
             var funcion = new Funcion
             {
                 SalaId = funcionDTO.SalaId,
-                Fecha = DateTime.Parse(funcionDTO.Fecha.ToString("yyyy-MM-dd")),
-                Horario = DateTime.Parse(funcionDTO.Horario.ToString("HH:mm:ss")),
+                Fecha = funcionDTO.Fecha,
+                Horario = funcionDTO.Horario,
                 PeliculaId = funcionDTO.PeliculaId
-                // Asigna otros campos según corresponda.
             };
 
-            // Agrega la nueva Funcion a la base de datos.
             _context.Funciones.Add(funcion);
             await _context.SaveChangesAsync();
 
-            // Puedes mapear la entidad Funcion nuevamente a FuncionDTO si es necesario.
-            // Esto dependerá de tu lógica de negocio y de si deseas devolver algún dato adicional.
+            // Resto del código para crear la respuesta...
 
             return CreatedAtAction("GetFuncion", new { id = funcion.FuncionId }, funcionDTO);
         }
+
+
+        //[HttpPost]
+        //public async Task<ActionResult<FuncionDTO>> Post2Funcion(FuncionDTO funcionDTO)
+        //{
+        //    if (funcionDTO == null)
+        //    {
+        //        return BadRequest("El objeto funcionDTO es nulo.");
+        //    }
+
+        //    // Mapea los campos de funcionDTO a la entidad Funcion.
+        //    var funcion = new Funcion
+        //    {
+        //        SalaId = funcionDTO.SalaId,
+        //        Fecha = DateTime.Parse(funcionDTO.Fecha.ToString("yyyy-MM-dd")),
+        //        Horario = DateTime.Parse(funcionDTO.Horario.ToString("HH:mm:ss")),
+        //        PeliculaId = funcionDTO.PeliculaId
+        //    };
+
+        //    _context.Funciones.Add(funcion);
+        //    await _context.SaveChangesAsync();
+
+
+        //    return CreatedAtAction("GetFuncion", new { id = funcion.FuncionId }, funcionDTO);
+        //}
 
         // DELETE: api/Funciones/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteFuncion(int id)
         {
             var funcion = await _context.Funciones.FindAsync(id);
+
             if (funcion == null)
             {
                 return NotFound();
+            }
+
+            // Verificar si hay tickets vendidos para esta función
+            var ticketsVendidos = await _context.Tickets.AnyAsync(t => t.FuncionId == id);
+
+            if (ticketsVendidos)
+            {
+                return BadRequest("No se puede eliminar la función porque tiene tickets vendidos.");
             }
 
             _context.Funciones.Remove(funcion);
@@ -154,6 +189,7 @@ namespace CinemaWebAPI.Controllers
 
             return NoContent();
         }
+
 
         private bool FuncionExists(int id)
         {
